Separate creation of sensors from rest of logic
fix CMakeLists

Remove unneeded transform

separate assignment of transform to sensor from rest
