+++ src/creo2urdf/CMakeLists.txt
@@ -9,11 +9,13 @@ add_library(creo2urdf::creo2urdf ALIAS creo2urdf)
 
 set(CREO2URDF_HDRS include/creo2urdf/Creo2Urdf.h
                    include/creo2urdf/Validator.h
+                   include/creo2urdf/Sensorizer.h
                    include/creo2urdf/Utils.h
 )
 set(CREO2URDF_SRCS src/main.cpp
                    src/Creo2Urdf.cpp
                    src/Validator.cpp
+                   src/Sensorizer.cpp
                    src/Utils.cpp
 )
